requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
compressai==1.2.6
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
einops==0.8.0
filelock==3.17.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
grpcio==1.71.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==6.30.2
psutil==6.1.1
py-cpuinfo==9.0.0
pycocotools==2.0.8
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.6.0
torch-geometric==2.6.1
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.72
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
